lil upd likes

+++ app/models/like.rb
@@ -3,12 +3,19 @@ class Like < ApplicationRecord
     where(post_id: post_id).count
   end
 
+  def self.arePostLikedBy(post_id, user_id)
+    where(post_id: post_id, user_id: user_id).count >= 1
+  end
+
   def self.like(post_id, user_id)
-    @like = Like.new(post_id: post_id, user_id: user_id)
-    @like.save
+    return if arePostLikedBy post_id, user_id
+
+    Like.new(post_id: post_id, user_id: user_id).save
   end
 
   def self.dislike(post_id, user_id)
+    return unless arePostLikedBy post_id, user_id
+
     Like.delete_by(post_id: post_id, user_id: user_id)
   end
 end
